Implement basic Single Sign On using LTI.
This is a really rough working version of sso authentication using LTI. Most of the work is handled by the new Authen::LTI module.

Note that because LTI authentication uses POST instead of GET, it creates a problem where if courses just use the root URL for Webwork, it can't get a course context. Previously in Vista SSO, I just redirected with all the GET parameters to get a course context. It seems the equivalent redirect with POST is a bit complicated, so I have left that for later. Currently, it's using a workaround where the moodle course just uses the Webwork URL with the course. A possible fix would be to store the POST in a cookie.

The OAuth message signing part of the LTI spec requires the installation of the Perl module Net::OAuth.

I'm not sure if Net::OAuth deals with nonces correctly, it doesn't appear to use any kind of persistent storage, so nonces might not be checked for uniqueness. This is partially mitigated by the timestamps though. Still a possible security issue.

There are corresponding changes to global.conf, namely, the addition of a the new Authen::LTI module.