More robust SSO LTI login.
The previous implementation of SSO login requires the course id to be in
the Webwork URL being accessed, it now works with just the webwork root
url.

The course id can only be added through redirection. The problem with
redirection is that it only supports GET parameters. Unfortunately, LTI
uses POST. The current workaround is to just convert all the LTI POST
parameters to GET and redirect to that URL.

Note that the warning messages about cookies in Authen.pm still exists
for some reason, not sure why.

Also added was a base for the course import wizard yet to be
implemented.