New architecture, class imports now working.
The goal of this new architecture is to better delineate reusable code
from platform specific code.

Platform specific code goes in Bridges. The Bridge is basically all the
custom code that used to be in WeBWorK.pm.

The BridgeManager tries all available Bridges to see if we have a
compatible Bridge to use according to the parameters passed. It then
uses that compatible Bridge instance to create/update courses or run
single sign on.

The Parser modules that goes with each Bridge implementation converts
the data received from whatever LMS it's talking to into a standard
format. This format is a hash containing course information and an array
containing student information.

Course information:
profid - the id of the professor teaching this course
name - the course's name, this is what will be used by Webwork to
identify the course
title - the long title of the course
id - the id used by the LMS to identify this course

Student information:
firstname
lastname
loginid - the id used by Webwork to identify this student
studentid - the id of the student assigned by the school
email
password - if left blank, password is just the student's loginid

The Importer modules uses the data obtained from the LMS to either
create or update a course. Note the Error module is here cause the
Importer modules are heavy users of it. It's not exactly the best error
system, but it'll have to do for now.

Might be a good idea to also move all custom ContentGenerator and Authen
modules into here.

Still need to reduce the amount of hard coding done, especially in
regard to OAuth secrets.