Display success or error messages on course import.
The LTI specific ContentGenerator has been modified to be the general course import status display. The original plan was to have LTI import be a step that requires user interaction. However, that proved to be a bit difficult to implement, and so was dropped. With user interaction dropped, the LTI display module behaved exactly like the Vista so, and so I just generalized it.

Errors are stored in Apache::Request's note hash under 'import_error'.