Now working with Vista.
Ported the Vista login and import functionality over as bridges.

Unfortunately, had to change the course name sanitization system because
of Vista's rather unfriendly course name handling. The function
sanitizeCourseName has now been pushed into the generic Parser class.
Might have to give parsers their own folder now since they now have a
generic class.

CourseUpdater now checks to see if the course environment was
initialized with the course name. If the ce does not have a course name,
the database object we get from it would be useless for getting class
information. We create a correctly initialized ce if the course name was
not found.