Fix Vista SSO login.
Vista SSO login depends on some config values stored in global.conf.
Unfortunately, these config values were renamed in the conversion to this bridge
architecture. The authentication module was still trying to read the old values.

The data import from Vista did not include the prof's information, so the prof
account cannot login since the account does not exist. This was fixed by adding
the prof info the the list of people to be imported.