Fix LTI OAuth security, class id in roster request.
There was a hole in the OAuth component of LTI where the message hash
wasn't being verified if a course hasn't been created. Previously, the
OAuth verification only took place when a user tries to log in. It did
not take place before roster requests. Now fixed so that OAuth
verification takes place before roster requests. Unfortunately, this
results in a slight duplication as if a course exists, not only is the
course updated but the user tries to login. This results in the message
hash being verified twice. This shouldn't be an issue later once we
restrict when course updates take place.

The class id in LTI roster requests were being taken from the parameters
for setting class grades instead of the parameters for rosters. This was
fixed.

Also a minor fix to explicitly make sure createCourse and updateCourse
of the Bridge class returns 0 if everything went ok.