If no section number, default to section string.
Courses sometimes might not have a section number, but a section title. In this
case, the previous behaviour was to just drop the section altogether. The
current behaviour will just add the section title on.

Note that LTI doesn't have a section parameter, so we still use no section for
that case.