Add logging of Vista course name to loginlist.
Webworks has limitations on the course name that Vista doesn't, so there is a
conversion from Vista's course name to Webworks. This mapping is now recorded
as part of loginlist.

Also, _getCourseName was working but emitting an error message because it
wasn't sure about what string to use the section number regex on. This was
fixed.