Parser objects now need the Apache request variable.
Webwork stores a lot of data in the Apache request variable $r. It turns out
that I now need to read global.conf values in the VistaParser, might as well
make it so that Parser objects need to take the $r variable in, since I might
need it later elsewhere too.

Constructors for VistaParser and LTIParser were changed to call the super
constructor instead of doing the construction themselves.

Noticed that LTIParser wasn't inheriting from Parser, fixed.