Assign homework to newly imported students.
When a new student is imported into the course, they are not assigned
homeworks by default. This turns out to be a bit troublesome, as we
have to go in manually to assign these new students the homework.

This patch now automatically assigns all student visible homework
sets to new users.