Add email info from text file.
This only takes effect when creating a new course. Since Vista does not
provide email address information, we can put together a text file
with the information and tell Webworks to use it.

The text file is named emaillist and takes the follow format:
<id number><tab><email address>

The text file is pointed to by a variable in global.conf called
$bridge{emaillist}