Remove dependency on PG for autoupdate.
The last change for automatically assigning homeworks to new students
brought in a dependency on PG because of the need to import
WeBWorK::ContentGenerator::Instructor. ContentGenerator actually also
brings in a stack of other dependencies, PG is just the first one that
it errors out on. Since the functions we need from Instructor doesn't
need all that dependency, it only needs a db object, it was easier to
just copy the functions into the CourseUpdater object instead of trying
to figure out how to satisfy all the dependencies.